+++ AGHRacingLogger/SW4STM32/AGHRacingLogger/Application/User/current_data_provider.h
@@ -5,6 +5,6 @@
 
 void saveCurrentData(uint8_t channel, uint16_t value);
 volatile uint16_t* getCurrentData();
-volatile uint32_t getCurrentDataForChannel(uint8_t channel);
+volatile uint16_t getCurrentDataForChannel(uint8_t channel);
 
 #endif /* __current_data_provider_H */
